move update.state into its own file
